Agregado manejo de errores varios

+++ src/routes/api/cartsRouter.js
@@ -1,9 +1,11 @@
 import { Router } from "express";
 import CartsManagerMongo from "../../daos/MongoDb/carts.manager.mongo.js";
+import ProductsManagerMongo from "../../daos/MongoDb/products.manager.mongo.js";
 
 const router = Router();
 
 const cartService = new CartsManagerMongo();
+const productService = new ProductsManagerMongo();
 
 router.get('/', async (req, res) => {
     try {
@@ -21,10 +23,14 @@ router.put('/:cid', async (req, res) => {
     try {
         const {body} = req;
         const {cid} = req.params;
+        if(!await cartService.isValidId(cid)) return res.status(400).send({status: 'error', message: 'El Id del carrito no tiene un formato valido'});
+
         const cart = await cartService.getCart({_id: cid});
 
         if(!cart) return res.status(404).send({status: 'error', message: 'No se encontro un carrito con este id'});
-        
+        if(!await productService.isValidId(body.product)) return res.status(400).send({status: 'error', message: 'El Id del producto no tiene un formato valido'});
+        if(body.quantity <= 0) return res.status(400).send({status: 'error', message: 'La cantidad agregada debe ser mayor que 0'});
+
         const addedProduct = await cartService.addProductToCart(cid, body);
         
         return res.status(200).send({status: 'success', payload: addedProduct});
@@ -42,9 +48,13 @@ router.put('/:cid/products/:pid', async (req, res) => {
 
         body = {quantity: Number(body.quantity)};
         const {cid, pid} = req.params;
+        if(!await cartService.isValidId(cid)) return res.status(400).send({status: 'error', message: 'El Id del carrito no tiene un formato valido'});
+        if(!await cartService.isValidId(pid)) return res.status(400).send({status: 'error', message: 'El Id del producto no tiene un formato valido'});
+
         const cart = await cartService.getCart({_id: cid});
 
         if(!cart) return res.status(404).send({status: 'error', message: 'No se encontro un carrito con este id'});
+        if(!await cartService.isProductOnCart(cid, pid)) return res.status(404).send({status: 'error', message: 'No se encontro un producto con este id en el carrito'});
         
         const addedProduct = await cartService.updateProductOnCart(cid, pid, body);
         
@@ -59,13 +69,15 @@ router.put('/:cid/products/:pid', async (req, res) => {
 router.delete('/:cid', async (req, res) => {
     try {
         const {cid} = req.params;
+        if(!await cartService.isValidId(cid)) return res.status(400).send({status: 'error', message: 'El Id del carrito no tiene un formato valido'});
+
         const cart = await cartService.getCart({_id: cid});
         
         if(!cart) return res.status(404).send({status: 'error', message: 'No se encontro un carrito con este id'});
         
-        const deltedCart = await cartService.deleteAllProducts(cid);
+        const deletedCart = await cartService.deleteAllProducts(cid);
         
-        return res.status(200).send({status: 'success', payload: deltedCart});
+        return res.status(200).send({status: 'success', payload: deletedCart});
         
     } catch (error) {
         console.log(error);
@@ -76,9 +88,12 @@ router.delete('/:cid', async (req, res) => {
 router.delete('/:cid/products/:pid', async (req, res) => {
     try {
         const {cid, pid} = req.params;
+        if(!await cartService.isValidId(cid)) return res.status(400).send({status: 'error', message: 'El Id del carrito no tiene un formato valido'});
         const cart = await cartService.getCart({_id: cid});
         
+        if(!await cartService.isValidId(pid)) return res.status(400).send({status: 'error', message: 'El Id del producto no tiene un formato valido'});
         if(!cart) return res.status(404).send({status: 'error', message: 'No se encontro un carrito con este id'});
+        if(!await cartService.isProductOnCart(cid, pid)) return res.status(404).send({status: 'error', message: 'No se encontro un producto con este Id en el carrito indicado'});
         
         const addedProduct = await cartService.deleteProductOnCart(cid, pid);
